Things are still breaking as we move from config to config.
I want to keep impossible state impossible, but for some of the Selection mode transitions I need to know what the current state of the DOM is.

We _could_ pass everything and kitchen sink in every time a DOM node changes but I think it would be more elegant to avoid that.