Fixing the bug that was causing custom selected options to end up with the selected value index of 0.
To fix this I had to change up some of the logic and that broke some unit tests
having to do with merging options. I think I solved by by making the `selectOptions`
function a bit more complicated. It's behavior seems pretty sensible so I think it's OK.

But I can't help but wonder if the options that are coming in to be selected should already have a selected index on them.